Refactored duotimer to add hooks to decrease component size
